requirements.txt
ansible-core
docker
jmespath
python-dateutil
requests
